refactor: 리뷰반영 및 디렉토리 구조변경
