+++ src/js/utils/dom.js
@@ -1,0 +1,3 @@
+export const $ = selector => document.querySelector(selector);
+
+export const $$ = selectors => document.querySelectorAll(selectors);
